+++ src/main/java/org/morphix/convert/ObjectConverter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 the original author or authors.
+ * Copyright 2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -40,8 +40,7 @@ public class ObjectConverter<S, D> implements
 		InstanceFunction<D>,
 		InstanceConvertFunction<S, D>,
 		SimpleConverter<S, D>,
-		ConvertFunction<S, D>,
-		Serializable {
+		ConvertFunction<S, D> {
 
 	/**
 	 * Serial version UID, required for {@link Serializable} classes.